Address review: stale slot indices, cutoff coverage, and a benchmark regression
Three things review turned up on the first two commits.

**A silent entry-loss path the abort fix did not close.** CodeRabbit and my own
reading independently landed on the same gap: `findSlot`/`findKey` derive a
mask and choose a slot, but a `.custom` equality procedure is arbitrary Scheme
and one that inserts into the table being probed makes `rehash` install a new,
larger `entries`. There is no use-after-free -- `ht.entries` is re-read -- but
the chosen index names an arbitrary bucket in the new layout, so the caller's
write lands on a live entry for a different key while `count` is still
incremented. One pre-existing key destroyed, `count` no longer matching the
live set. A (prefill x inject x span) sweep reproduced it in **128 of 210**
combinations; all 210 are clean now. The probe restarts when it observes the
layout move, bounded at 16 restarts so a procedure that mutates on every call
raises instead of spinning.

**The new #2023 assertions did not reach either cutoff.** `mk` builds a flat
list, and after the fix a flat list no longer spends the nesting budget -- so
nothing exercised `DEEP_CUTOFF_HASH` or the `MAX_HASH_SPINE` truncation. Added
a `nest` helper for keys deeper than `MAX_HASH_DEPTH`, and a case whose only
distinguishing element sits past the spine cutoff so every key hashes alike and
`deepEqual` has to separate them. The duplicate 200-deep-keys cell became the
nesting one rather than a second copy of a test three lines away.

**A 1.32x regression on `benchmarks/hashtable.scm`.** Routing the default
`.equal` mode through `hashForTable`/`equalForTable` -- whose `.custom` arms
pull in the whole VM-call path -- stopped `valueHash` and `deepEqual` from
inlining: 18% on insert, 10% on lookup. The `.equal` specialization the
original code had is restored, and the other callback-free modes get a probe
loop without the restart bookkeeping. `HashEntry.hash` is also `u32` rather
than `usize`: only `hash & (capacity - 1)` is ever read, and it keeps the
struct at its original 24 bytes. Now 0.0302s against 0.0306s on main.

Also: strengthened the #2024 abort assertion, which accepted entry loss
(`(>= size 5)` passes while a key is missing); corrected an inaccurate test
name (an f64vector and a bytevector with the same elements are not the same
byte length); and made the deep-copy test assert the copied entry's cached hash
agrees with the *destination* table -- not with the source, since the
non-custom arm deliberately re-hashes as it re-buckets.

+++ src/types_hashtable.zig
@@ -20,7 +20,13 @@ pub const HashEntry = struct {
     /// procedure from running while a soon-to-be-freed entry array is live
     /// (#2024). Deliberately has no default so every insertion site must
     /// supply the hash `findSlot` already computed.
-    hash: usize,
+    ///
+    /// `u32`, not `usize`: only `hash & (capacity - 1)` is ever read, and a
+    /// table with 2^32 buckets would need ~96 GB of entries, so the high half
+    /// can never select a bucket. It also keeps this struct at 24 bytes --
+    /// widening it to 32 cost 17% on the `hashtable` benchmark, which probes
+    /// far more entries than it stores.
+    hash: u32,
     state: HashEntryState = .empty,
 };
